# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for ecology section of arctech DATA",
        "title": "SunTechnics API - OpenAPI",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/deposits": {
            "get": {
                "description": "Получить список месторождений в заданном регионе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deposits"
                ],
                "summary": "Получить список месторождений в заданном регионе",
                "operationId": "get-deposits",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "Id региона",
                        "name": "regionId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getDepositsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/minerals": {
            "get": {
                "description": "Получить список полезных ископаемых указанного типа в заданном регионе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "minerals"
                ],
                "summary": "Получить список полезных ископаемых указанного типа в заданном регионе",
                "operationId": "get-minerals",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "ID региона",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "ID типа полезных ископаемых",
                        "name": "typeId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getMineralsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/mineraltypes": {
            "get": {
                "description": "Получить список типов полезных ископаемых",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mineral types"
                ],
                "summary": "Получить список типов полезных ископаемых",
                "operationId": "get-mineral-types",
                "parameters": [
                    {
                        "type": "string",
                        "format": "int",
                        "description": "ID региона",
                        "name": "regionId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getMineraltypesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/regions": {
            "get": {
                "description": "Получить список наблюдаемых регионов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regions"
                ],
                "summary": "Получить список наблюдаемых регионов",
                "operationId": "get-all-regions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getAllRegionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/response": {
            "get": {
                "description": "Получить расчет по мониторингу ММП",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "response"
                ],
                "summary": "Получить расчет по мониторингу ММП",
                "operationId": "get-response",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "ID of the region where minerals are extracted",
                        "name": "regionId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "ID of the mineral",
                        "name": "mineralId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getResponseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.getAllRegionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Region"
                    }
                }
            }
        },
        "handler.getDepositsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Deposit"
                    }
                }
            }
        },
        "handler.getMineralsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Minerals"
                    }
                }
            }
        },
        "handler.getMineraltypesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MineralTypes"
                    }
                }
            }
        },
        "handler.getResponseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Response"
                    }
                }
            }
        },
        "models.Deposit": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.MineralTypes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Minerals": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Region": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "coast_dist": {
                    "type": "number"
                },
                "coast_rate": {
                    "type": "number"
                },
                "deposit_name": {
                    "type": "string"
                },
                "g_temp": {
                    "type": "number"
                },
                "oopt": {
                    "type": "string"
                },
                "perm_distr": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "t_trend": {
                    "type": "number"
                },
                "yed_dist": {
                    "type": "number"
                }
            }
        }
    }
}